moved exportCounts into separate class
